Remove framework, lens, search, and state modules
