+++ crates/dock-alloc-solver/src/lib.rs
@@ -20,9 +20,6 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 pub mod domain;
-pub mod framework;
-pub mod lens;
+pub mod ledger;
 pub mod occ;
 pub mod quay;
-pub mod search;
-pub mod state;
